fix move the util function to action folder
